fix: fixed table migrations
