--- alembic/versions/588e8fa12b62_.py
@@ -1,60 +1,0 @@
-"""empty message
-
-Revision ID: 588e8fa12b62
-Revises: 7c93756cfcd7
-Create Date: 2025-07-04 12:40:43.139163
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects.postgresql import VARCHAR
-
-
-
-# revision identifiers, used by Alembic.
-revision: str = '588e8fa12b62'
-down_revision: Union[str, None] = '7c93756cfcd7'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('answers',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=True),
-    sa.Column('quiz_id', sa.Integer(), nullable=False),
-    sa.Column('question_id', sa.Integer(), nullable=False),
-    sa.Column('submitted_answer', sa.Text(), nullable=False),
-    sa.Column('is_correct', sa.Boolean(), nullable=False),
-    sa.Column('timestamp', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['question_id'], ['questions.id'], ),
-    sa.ForeignKeyConstraint(['quiz_id'], ['quizzes.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_answers_id'), 'answers', ['id'], unique=False)
-    op.add_column('documents', sa.Column('document_id', sa.String(), nullable=True))
-    op.create_index(op.f('ix_documents_document_id'), 'documents', ['document_id'], unique=True)
-    op.alter_column('questions', 'difficulty',
-               existing_type=sa.VARCHAR(length=6),
-               type_=VARCHAR(6),
-               existing_nullable=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('questions', 'difficulty',
-               existing_type=VARCHAR(6),
-               type_=sa.VARCHAR(length=6),
-               existing_nullable=False)
-    op.drop_index(op.f('ix_documents_document_id'), table_name='documents')
-    op.drop_column('documents', 'document_id')
-    op.drop_index(op.f('ix_answers_id'), table_name='answers')
-    op.drop_table('answers')
-    # ### end Alembic commands ###
