+++ alembic/versions/d4664eb8ea67_add_answers_table.py
@@ -1,0 +1,75 @@
+"""add_answers_table
+
+Revision ID: d4664eb8ea67
+Revises: d7be7699a9f9
+Create Date: 2025-07-05 21:24:15.758252
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'd4664eb8ea67'
+down_revision: Union[str, None] = 'd7be7699a9f9'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - adjusted for NOT NULL ###
+    op.create_table(
+        'answers',
+        sa.Column('id', sa.Integer(), nullable=False),
+        sa.Column('user_id', sa.Integer(), nullable=True),
+        sa.Column('quiz_id', sa.Integer(), nullable=False),
+        sa.Column('question_id', sa.Integer(), nullable=False),
+        sa.Column('submitted_answer', sa.Text(), nullable=False),
+        sa.Column('is_correct', sa.Boolean(), nullable=False),
+        sa.Column('timestamp', sa.DateTime(), nullable=True),
+        sa.ForeignKeyConstraint(['question_id'], ['questions.id']),
+        sa.ForeignKeyConstraint(['quiz_id'], ['quizzes.id']),
+        sa.ForeignKeyConstraint(['user_id'], ['users.id']),
+        sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_answers_id'), 'answers', ['id'], unique=False)
+
+    op.add_column('documents', sa.Column('document_id', sa.String(), nullable=True))
+    op.create_index(op.f('ix_documents_document_id'), 'documents', ['document_id'], unique=True)
+
+    # Add column with temporary default
+    op.add_column('questions', sa.Column('difficulty', sa.String(6), nullable=False, server_default='easy'))
+
+    # Remove the default after column has been populated
+    op.alter_column('questions', 'difficulty', server_default=None)
+
+    # Update type column if needed
+    op.alter_column(
+        'questions',
+        'type',
+        existing_type=sa.String(length=7),
+        type_=sa.String(7),
+        existing_nullable=False
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic ###
+    op.alter_column(
+        'questions',
+        'type',
+        existing_type=sa.String(7),
+        type_=sa.String(length=7),
+        existing_nullable=False
+    )
+    op.drop_column('questions', 'difficulty')
+    op.drop_index(op.f('ix_documents_document_id'), table_name='documents')
+    op.drop_column('documents', 'document_id')
+    op.drop_index(op.f('ix_answers_id'), table_name='answers')
+    op.drop_table('answers')
+    # ### end Alembic commands ###
